backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-dotenv==1.0.0
openai==0.28.0
pydantic==2.3.0